refactor(webapp): address review on the overdue alerts
- Move OverdueInvoiceAlert out of components/patterns, which is for generic
  reusable pieces, into features/Billing. Both the sidebar and the Billing page
  reach it there without layout depending on pages.
- Cut the comments back. Most restated the code; the rest were far longer than
  what they explained.
- Drop the claim that switching to a free plan takes effect in-session. Plan
  changes wait for the end of the billing period.
- Shorten the overdue cache from 12h to 10min, so the alert clears soon after an
  invoice is paid rather than lingering for the rest of the day.
- Rename showOverdueCard to showOverdueAlert, matching showUsageAlert.
